Add files via upload
Simple function definition to pull current time and write/append to file. Will be ultimately used for implementation in an auto-open feature of the virtual assistant. The virtual assistant will initially be manually called, but will eventually learn ideal times to open, queues for opening and dismissal, and develop appropriate functionality accordingly.